+++ php/20180418-19/uzdavinys1-2.php
@@ -1,0 +1,17 @@
+    <?php
+    class Mokinys extends Trimestras {
+        
+        public $vardas;
+        public $pavarde;
+
+        public function __construct ($vardas, $pavarde, $trimestras, $vidurkis) {
+            parent:: __construct ($trimestras, $vidurkis);
+            $this->vardas = $vardas;
+            $this->pavarde = $pavarde;
+        }
+
+        public function pilnasVardas() {
+            return $this->vardas . ' ' . $this->pavarde;
+        }
+    }
+?>
