+++ php/20180418-19/uzdavinys1-3.php
@@ -1,0 +1,10 @@
+    <?php
+$eustarchijus = new Mokinys ('Eustarchijus', 'Vamzdys', ['lietuviu' => 8, 'matematika' => 8, 'anglu' => 8], 0);
+$kulverstukas = new Mokinys ('Kūlverstukas', 'Didžiaausis', ['lietuviu' => 10, 'matematika' => 9, 'anglu' => 8], 0);
+$paslemekas = new Mokinys ('Pašlemėkas', 'Šliūpšta', ['lietuviu' => 5, 'matematika' => 9, 'anglu' => 7], 0);
+$bonifacijus = new Mokinys ('Liūtas', 'Bonifacijus', ['lietuviu' => 10, 'matematika' => 10, 'anglu' => 10], 0);
+$perestukinas = new Mokinys ('Perestukinas', 'Nemokusis', ['lietuviu' => 5, 'matematika' => 3, 'anglu' => 4], 0);
+$leopoldas = new Mokinys ('Katinas', 'Leopoldas', ['lietuviu' => 8, 'matematika' => 4, 'anglu' => 6], 0);
+
+$mokiniai = [$eustarchijus, $kulverstukas, $paslemekas, $bonifacijus, $perestukinas, $leopoldas];
+?>
